// frontend/src/backend.d.ts
export interface Location {
    latitude: number;
    longitude: number;
    timestamp: Time;
}
export type Time = bigint;
// ...
export interface StaffAttendance {
    id: string;
    staffId: string;
    date: Time;
    hoursWorked: bigint;
    clockOut?: Time;
    clockIn: Time;
    totalPay: bigint;
    locationHistory: Array<Location>;
    payRate: bigint;
    location?: Location;
}